+++ src/main/java/com/example/controledefinancas/repository/FinancaRepository.java
@@ -4,6 +4,6 @@ import com.example.controledefinancas.model.Financa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface FinancaRepository extends JpaRepository<Financa, Long> {
+public interface FinancaRepository extends JpaRepository<Financa, Integer> {
 
 }
